fix: moving header, footer file directory and modified header size
